--- epl_scrapy/epl_scrapy/spiders/arsenal_goal_2022_23_spider.py
@@ -1,9 +1,0 @@
-from base import BaseGoalSpider
-from epl_scrapy.match_ids.season_2023 import match_ids
-
-BASE_MATCH_URL = "https://www.premierleague.com/match/"
-
-class Arsenal2023GoalSpider(BaseGoalSpider):
-    """Goals of Arsenal in season 2022/23."""
-
-    start_urls = ["".join([BASE_MATCH_URL, match_id]) for match_id in match_ids]
